Fix inl/outl operand type mismatch in ports.h
inl() and outl() declared their port parameter as uint32_t, but the x86
in/out instructions only accept a 16-bit port number (DX register, not
EDX). With newer binutils, gcc emits 'inl %edx, %eax' which the
assembler rejects with:

  include/../common/ports.h:34: Error: operand type mismatch for 'out'
  include/../common/ports.h:29: Error: operand type mismatch for 'in'

The other port I/O functions (inb/outb/inw/outw) all correctly use
uint16_t port. This commit makes inl/outl consistent.

x86 I/O ports are 16-bit addressed, so the type change is semantically
correct. Callers that pass uint32_t values (e.g. io_base + offset
arithmetic) undergo an implicit narrowing conversion which is silent
under default -Wall.

Verified: pci.c, net_rtl8139.c, net_e1000.c, ahci.c all compile cleanly
with -Wall -Wno-unused-parameter and produce no errors.

+++ common/ports.h
@@ -24,13 +24,13 @@ static inline void outw(uint16_t port, uint16_t data) {
     asm volatile("outw %0, %1" : : "a"(data), "d"(port));
 }
 
-static inline uint32_t inl(uint32_t port) {
+static inline uint32_t inl(uint16_t port) {
     uint32_t result;
     asm volatile("inl %1, %0" : "=a"(result) : "d"(port));
     return result;
 }
 
-static inline void outl(uint32_t port, uint32_t data) {
+static inline void outl(uint16_t port, uint32_t data) {
     asm volatile("outl %0, %1" : : "a"(data), "d"(port));
 }
